pyproject.toml
[project]
name = "bonsainet"
version = "0.1.0"
description = ""
authors = [
    {name = "Ayoub G.",email = "dev@ayghri.me"}
]
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "torch (>=2.9.1,<3.0.0)",
    "torchvision (>=0.24.1,<0.25.0)",
    "hydra-core (>=1.3.2,<2.0.0)",
    "wandb (>=0.23.1,<0.24.0)",
    "transformers (>=4.57.3,<5.0.0)",
    "datasets (>=4.4.1,<5.0.0)",
    "accelerate (>=1.12.0,<2.0.0)",
    "matplotlib (>=3.10.7,<4.0.0)",
    "lm-eval (>=0.4.9.2,<0.5.0.0)"
]

[tool.poetry]
packages = [{include = "bonsainet", from = "./"}]


[tool.poetry.group.dev.dependencies]
jupyter = "^1.1.1"
ipywidgets = "^8.1.8"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
